some code refactor: `config::MPVClient` and `config::ReqClient`
