+++ src/ui/new_dropsel.rs
@@ -1,3 +1,4 @@
+use super::config::MPVClient;
 use super::network;
 use gtk::prelude::*;
 use gtk::Orientation;
@@ -94,7 +95,7 @@ pub fn newmediadropsel(playbackinfo: network::Media, id: String) -> gtk::Box {
                     network::runtime().spawn(async move {
                         let _ = network::markwatched(name, sourceid).await;
                     });
-                    network::mpv_play(directurl.expect("no url"), media.Name.clone());
+                    MPVClient::play(directurl.expect("no url"), media.Name.clone());
                 }
             }
             return;
@@ -115,7 +116,7 @@ pub fn newmediadropsel(playbackinfo: network::Media, id: String) -> gtk::Box {
                                 network::runtime().spawn(async move {
                                     let _ = network::markwatched(name, sourceid).await;
                                 });
-                                let _ = network::mpv_play_withsub(
+                                let _ = MPVClient::play_with_sub(
                                     directurl.expect("no url"),
                                     suburl.expect("no url"),
                                     media.Name.clone(),
@@ -126,10 +127,8 @@ pub fn newmediadropsel(playbackinfo: network::Media, id: String) -> gtk::Box {
                                 network::runtime().spawn(async move {
                                     let _ = network::markwatched(name, sourceid).await;
                                 });
-                                let _ = network::mpv_play(
-                                    directurl.expect("no url"),
-                                    media.Name.clone(),
-                                );
+                                let _ =
+                                    MPVClient::play(directurl.expect("no url"), media.Name.clone());
                             }
                         }
                     }
